pyproject.toml
[project]
name = "quiz-master"
version = "0.1.0"
description = "An interactive CLI to create home server applications as Helm Charts"
readme = "README.md"
dependencies = [
    "questionary>=2.1.0",
    "pyinstaller (>=6.12.0,<7.0.0)",
    "jinja2>=3.1.6",
    "ruff>=0.13.0",
    "pyyaml>=6.0.3",
]
